requirements.txt
invocate>=0.2.1

psycopg2-binary==2.9.9
torch>=2.0.0
transformers>=4.21.0
numpy>=1.21.0
vllm>=0.2.0
sentence-transformers>=2.2.0
attrs
pyyaml
